PicoVM small steps ahead...
